refactor: rename main package to epam
