+++ settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'epam_web'
+rootProject.name = 'epam'
