+++ src/main/java/grvt/cloud/epam/EpamWebApplication.java
@@ -1,6 +1,5 @@
-package grvt.cloud.epam_web;
+package grvt.cloud.epam;
 
-import io.swagger.v3.oas.annotations.OpenAPIDefinition;
 import io.swagger.v3.oas.models.OpenAPI;
 import io.swagger.v3.oas.models.info.Info;
 import org.springframework.boot.SpringApplication;
