+++ src/main/java/grvt/cloud/epam/perfomance_counter/Counter.java
@@ -1,0 +1,8 @@
+package grvt.cloud.epam.perfomance_counter;
+
+public class Counter extends Thread{
+    @Override
+    public synchronized void run() {
+        CounterLogic.Increment();
+    }
+}
